ANPR1/ANPR/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==0.25.3
azure-storage-blob
imutils==0.5.4
pytesseract==0.3.7
opencv-python==3.4.2.17
